Fix toneEndTime type

+++ dtmfService.h
@@ -14,7 +14,7 @@ class DtmfService
 {
     PwmOut *pwmOut1;
     PwmOut *pwmOut2;
-    uint toneEndTime;
+    unsigned long toneEndTime;
 
   public:
     DtmfService(PinName pin1Name, PinName pin2Name);
